+++ src/services/expenseService.ts
@@ -8,24 +8,17 @@ import {
   doc,
 } from "firebase/firestore";
 import { db } from "@/firebase";
-
-export type TExpense = {
-  id?: string;
-  billId: number;
-  description: string;
-  amount: string;
-  createdAt: string;
-};
+import { TExpense } from "@/ProjectType";
 
 const COLLECTION_NAME = "expenses";
 
 export const expenseService = {
   async createExpense(expense: TExpense): Promise<TExpense> {
-    const docRef = await addDoc(collection(db, COLLECTION_NAME), {
+    await addDoc(collection(db, COLLECTION_NAME), {
       ...expense,
       createdAt: new Date().toISOString(),
     });
-    return { ...expense, id: docRef.id };
+    return { ...expense };
   },
 
   async getExpensesByBillId(billId: number): Promise<TExpense[]> {
@@ -38,13 +31,16 @@ export const expenseService = {
       (doc) =>
         ({
           ...doc.data(),
-          id: doc.id,
         } as TExpense)
     );
   },
 
-  async deleteExpense(expenseId: string): Promise<void> {
-    const docRef = doc(db, COLLECTION_NAME, expenseId);
-    await deleteDoc(docRef);
+  async deleteExpenseById(id: number): Promise<void> {
+    const q = query(collection(db, COLLECTION_NAME), where("id", "==", id));
+    const querySnapshot = await getDocs(q);
+    if (!querySnapshot.empty) {
+      const docRef = doc(db, COLLECTION_NAME, querySnapshot.docs[0].id);
+      await deleteDoc(docRef);
+    }
   },
 };
